requirements.txt
asgiref==3.2.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.41.0
django-ckeditor==5.8.0
django-countries==5.5
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-import-export==2.0
django-jet==1.0.8
django-js-asset==1.2.2
django-storages==1.8
et-xmlfile==1.0.1
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-storage==1.24.1
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==20.0.4
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==3.0.2
Pillow==7.0.0
protobuf==3.11.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.7
xlrd==1.2.0
xlwt==1.3.0
